Module budget_executor::asynchronous::singlethreaded

A single-threaded (!Send + !Sync), asynchronous budgeting implementation that is runtime agnostic.

The only difference between this module and the threadsafe module is that this one uses std::rc::Rc and std::cell::RefCell instead of std::sync::Arc and std::sync::Mutex.

Structs

BudgetedFuture

Executes a future with a given budget when awaited.

Context

A budget-limited asynchronous context.

IncompleteFuture

A future that was budgeted using Context::run_with_budget() that has not yet completed.

WaitForBudgetFuture

A future that waits for additional budget to be allocated through ReplenishableBudget::replenish().

Enums

Progress

The progress of a future’s execution.