Module budget_executor::blocking

A standalone implementation does not require another async executor and blocks the current thread while executing.

Structs

IncompleteFuture

A future that was budgeted with run_with_budget() that has not yet completed.

Enums

Progress

The progress of a future’s execution.

Functions

run_with_budget

Executes future with the provided budget. The future will run until it completes or until it has invoked spend() enough to exhaust the budget provided. If the future never called spend(), this function will block the current thread until the future completes.