Module budget_executor::blocking::threadsafe

A threadsafe (Send + Sync), blocking budgeting implementation that is runtime agnostic.

The only difference between this module and the singlethreaded module is that this one uses std::sync::Arc and std::sync::Mutex instead of std::rc::Rc and std::cell::RefCell.

Structs

IncompleteFuture

A future that was budgeted with Runtime::run_with_budget() that has not yet completed.

Runtime

A lightweight asynchronous runtime that runs a future while keeping track of a budget.

TaskHandle

A handle to a task scheduled with a Runtime. Invoking .await on this type will return the task’s output when the original task completes.

Enums

Progress

The progress of a future’s execution.